Fixed G2 query negative timestamps. Made more verbose G2 query acknowledgment packet dumping.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7765 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/G2Neighbour.cpp
@@ -874,7 +874,7 @@ BOOL CG2Neighbour::ParseKHLPacket(CG2Packet* pPacket, SOCKADDR_IN* pHost)
 		G2_PACKET nType, nInnerType;
 		DWORD nLength, nInner;
 		BOOL bCompound;
-		DWORD tAdjust = ( pOwner ) ? pOwner->m_tAdjust : 0;
+		LONG tAdjust = ( pOwner ) ? pOwner->m_tAdjust : 0;
 		DWORD tNow = static_cast< DWORD >( time( NULL ) );
 
 		if ( pOwner && pOwner->m_pHubGroup ) pOwner->m_pHubGroup->Clear();
